+++ DMSport/Sources/Scene/Notice/UI/EntireNoticeVC.swift
@@ -79,22 +79,3 @@ class EntireNoticeVC: BaseVC {
         }
     }
 }
-
-extension EntireNoticeVC: UITableViewDataSource, UITableViewDelegate {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return noticeList.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        if let cell = tableView.dequeueReusableCell(withIdentifier: "EntireNotice", for: indexPath) as? NoticeCell {
-            cell.noticeTitle.text = "\(noticeList[indexPath.row].title)"
-            cell.noticeDetails.text = "\(noticeList[indexPath.row].created_at)"
-            cell.noticeContent.text = "\(noticeList[indexPath.row].content)"
-            
-            cell.selectionStyle = .none
-            return cell
-        } else {
-            return UITableViewCell()
-        }
-    }
-}
